requirements.txt
anthropic>=0.38.0
beautifulsoup4>=4.12.3
fastapi>=0.110.0
httpx>=0.27.0
pandas>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.8.0
pytest>=8.2.0
requests>=2.32.0
uvicorn[standard]>=0.30.0

